+++ engine/runtime-parser/src/main/java/org/enso/compiler/core/ir/expression/IfThenElse.java
@@ -1,5 +1,6 @@
 package org.enso.compiler.core.ir.expression;
 
+import org.enso.compiler.core.ir.Empty;
 import org.enso.compiler.core.ir.Expression;
 import org.enso.compiler.core.ir.IdentifiedLocation;
 import org.enso.compiler.core.ir.MetadataStorage;
@@ -10,6 +11,7 @@ import org.enso.runtime.parser.dsl.IRChild;
 /** Enso if/then(else) expression. */
 @GenerateIR(interfaces = {Expression.class})
 public final class IfThenElse extends IfThenElseGen {
+
   @GenerateFields
   public IfThenElse(
       @IRChild Expression condition,
@@ -38,6 +40,38 @@ public final class IfThenElse extends IfThenElseGen {
 
   public scala.Option<Expression> falseBranch() {
     return scala.Option.apply(falseBranchOrNull());
+  }
+
+  /**
+   * Builds "else only" instance.
+   *
+   * @param value the else branch
+   * @param where location of the value
+   * @param meta the metadaa in the tree
+   * @return insance of if/then/else
+   * @see #isOnlyElse()
+   */
+  public static IfThenElse buildOnlyElse(
+      Expression value, IdentifiedLocation where, MetadataStorage meta) {
+    var emptyBlock = Empty.builder().build();
+    return builder()
+        .condition(emptyBlock)
+        .trueBranch(emptyBlock)
+        .falseBranchOrNull(value)
+        .location(where)
+        .passData(meta)
+        .build();
+  }
+
+  /**
+   * Checks whether this object represents "pending" else branch. Such a branch only makes sense if
+   * merged to previous if statement.
+   *
+   * @return {@code true} if created by {@link #buildOnlyElse} factory
+   * @see #isOnlyElse()
+   */
+  public boolean isOnlyElse() {
+    return cond() instanceof Empty c && trueBranch() instanceof Empty t && c == t;
   }
 
   @Override