+++ CHANGELOG.md
@@ -8,11 +8,13 @@
 
 #### Enso Standard Library
 
+- [Multiline if_then_else][14522].
 - [Using dual JVM mode for Standard.Microsoft][14476].
 - [Standard.Test pending field is lazy][14536].
 - [Using dual JVM mode for Standard.AWS][14568].
 - [Polishing Standard.Test API][14599].
 
+[14522]: https://github.com/enso-org/enso/pull/14522
 [14476]: https://github.com/enso-org/enso/pull/14476
 [14536]: https://github.com/enso-org/enso/pull/14536
 [14568]: https://github.com/enso-org/enso/pull/14568